+++ datumaro/components/dataset.py
@@ -15,10 +15,13 @@ import inspect
 import logging as log
 import os
 import os.path as osp
+import warnings
 
 from datumaro.components.annotation import AnnotationType, LabelCategories
 from datumaro.components.config_model import Source
-from datumaro.components.converter import Converter
+from datumaro.components.converter import (
+    Converter, ExportContext, ExportErrorPolicy, _ExportFail,
+)
 from datumaro.components.dataset_filter import (
     XPathAnnotationsFilter, XPathDatasetFilter,
 )
@@ -30,9 +33,12 @@ from datumaro.components.errors import (
 )
 from datumaro.components.extractor import (
     DEFAULT_SUBSET_NAME, CategoriesInfo, DatasetItem, Extractor, IExtractor,
-    ItemTransform, Transform,
+    ImportContext, ImportErrorPolicy, ItemTransform, Transform, _ImportFail,
 )
 from datumaro.components.launcher import Launcher, ModelTransform
+from datumaro.components.progress_reporting import (
+    NullProgressReporter, ProgressReporter,
+)
 from datumaro.plugins.transforms import ProjectLabels
 from datumaro.util import is_method_redefined
 from datumaro.util.log_utils import logging_disabled
@@ -882,7 +888,9 @@ class Dataset(IDataset):
         self._data.flush_changes()
 
     @scoped
-    def export(self, save_dir: str, format: Union[str, Type[Converter]],
+    def export(self, save_dir: str, format: Union[str, Type[Converter]], *,
+            progress_reporter: Optional[ProgressReporter] = None,
+            error_policy: Optional[ExportErrorPolicy] = None,
             **kwargs) -> None:
         """
         Saves the dataset in some format.
@@ -892,7 +900,9 @@ class Dataset(IDataset):
             format - The desired output format.
                 If a string is passed, it is treated as a plugin name,
                 which is searched for in the dataset environment.
-            **kwargs - Parameters for the export format
+            progress_reporter - An object to report progress
+            error_policy - An object to report format-related errors
+            **kwargs - Parameters for the format
         """
 
         if not save_dir:
@@ -915,13 +925,59 @@ class Dataset(IDataset):
             inplace = False
         os.makedirs(save_dir, exist_ok=True)
 
-        if not inplace:
-            converter.convert(self, save_dir=save_dir, **kwargs)
-            if not self.is_bound:
-                self.bind(save_dir, format, options=copy(kwargs))
-                self.flush_changes()
-        else:
-            converter.patch(self, self.get_patch(), save_dir=save_dir, **kwargs)
+        has_ctx_args = progress_reporter is not None or error_policy is not None
+
+        if not progress_reporter:
+            progress_reporter = NullProgressReporter()
+
+        assert 'ctx' not in kwargs
+        converter_kwargs = copy(kwargs)
+        converter_kwargs['ctx'] = ExportContext(
+            progress_reporter=progress_reporter,
+            error_policy=error_policy)
+
+        try:
+            if not inplace:
+                try:
+                    converter.convert(self, save_dir=save_dir,
+                        **converter_kwargs)
+                except TypeError as e:
+                    # TODO: for backward compatibility. To be removed after 0.3
+                    if "unexpected keyword argument 'ctx'" not in str(e):
+                        raise
+
+                    if has_ctx_args:
+                        warnings.warn("It seems that '%s' converter "
+                            "does not support progress and error reporting, "
+                            "it will be disabled" % format,
+                            DeprecationWarning)
+                    converter_kwargs.pop('ctx')
+
+                    converter.convert(self, save_dir=save_dir,
+                        **converter_kwargs)
+            else:
+                try:
+                    converter.patch(self, self.get_patch(), save_dir=save_dir,
+                        **converter_kwargs)
+                except TypeError as e:
+                    # TODO: for backward compatibility. To be removed after 0.3
+                    if "unexpected keyword argument 'ctx'" not in str(e):
+                        raise
+
+                    if has_ctx_args:
+                        warnings.warn("It seems that '%s' converter "
+                            "does not support progress and error reporting, "
+                            "it will be disabled" % format,
+                            DeprecationWarning)
+                    converter_kwargs.pop('ctx')
+
+                    converter.patch(self, self.get_patch(), save_dir=save_dir,
+                        **converter_kwargs)
+        except _ExportFail as e:
+            raise e.__cause__
+
+        self.bind(save_dir, format, options=copy(kwargs))
+        self.flush_changes()
 
     def save(self, save_dir: Optional[str] = None, **kwargs) -> None:
         options = dict(self._options)
@@ -936,7 +992,28 @@ class Dataset(IDataset):
 
     @classmethod
     def import_from(cls, path: str, format: Optional[str] = None, *,
-            env: Optional[Environment] = None, **kwargs) -> Dataset:
+            env: Optional[Environment] = None,
+            progress_reporter: Optional[ProgressReporter] = None,
+            error_policy: Optional[ImportErrorPolicy] = None,
+            **kwargs) -> Dataset:
+        """
+        Creates a `Dataset` instance from a dataset on the disk.
+
+        Args:
+            path - The input file or directory path
+            format - Dataset format.
+                If a string is passed, it is treated as a plugin name,
+                which is searched for in the `env` plugin context.
+                If not set, will try to detect automatically,
+                using the `env` plugin context.
+            env - A plugin collection. If not set, the built-in plugins are used
+            progress_reporter - An object to report progress.
+                Implies earger loading.
+            error_policy - An object to report format-related errors.
+                Implies earger loading.
+            **kwargs - Parameters for the format
+        """
+
         if env is None:
             env = Environment()
 
@@ -955,17 +1032,59 @@ class Dataset(IDataset):
         else:
             raise UnknownFormatError(format)
 
-        extractors = []
-        for src_conf in detected_sources:
-            if not isinstance(src_conf, Source):
-                src_conf = Source(src_conf)
-            extractors.append(env.make_extractor(
-                src_conf.format, src_conf.url, **src_conf.options
-            ))
+        # TODO: probably, should not be available in lazy mode, because it
+        # becomes unreliable and error-prone. For progress reporting it
+        # makes little sense, because loading stage is spread over other
+        # operations. Error reporting is going to be unreliable.
+        has_ctx_args = progress_reporter is not None or error_policy is not None
+        eager = has_ctx_args
 
-        dataset = cls.from_extractors(*extractors, env=env)
+        if not progress_reporter:
+            progress_reporter = NullProgressReporter()
+        pbars = progress_reporter.split(len(detected_sources))
+
+        try:
+            extractors = []
+            for src_conf, pbar in zip(detected_sources, pbars):
+                if not isinstance(src_conf, Source):
+                    src_conf = Source(src_conf)
+
+                extractor_kwargs = dict(src_conf.options)
+
+                assert 'ctx' not in extractor_kwargs
+                extractor_kwargs['ctx'] = ImportContext(
+                    progress_reporter=pbar,
+                    error_policy=error_policy)
+
+                try:
+                    extractors.append(env.make_extractor(
+                        src_conf.format, src_conf.url, **extractor_kwargs
+                    ))
+                except TypeError as e:
+                    # TODO: for backward compatibility. To be removed after 0.3
+                    if "unexpected keyword argument 'ctx'" not in str(e):
+                        raise
+
+                    if has_ctx_args:
+                        warnings.warn("It seems that '%s' extractor "
+                            "does not support progress and error reporting, "
+                            "it will be disabled" % src_conf.format,
+                            DeprecationWarning)
+                    extractor_kwargs.pop('ctx')
+
+                    extractors.append(env.make_extractor(
+                        src_conf.format, src_conf.url, **extractor_kwargs
+                    ))
+
+            dataset = cls.from_extractors(*extractors, env=env)
+            if eager:
+                dataset.init_cache()
+        except _ImportFail as e:
+            raise e.__cause__
+
         dataset._source_path = path
         dataset._format = format
+
         return dataset
 
     @staticmethod