+++ datumaro/plugins/coco_format/converter.py
@@ -651,24 +651,35 @@ class CocoConverter(Converter):
         if self._save_dataset_meta:
             self._save_meta_file(self._save_dir)
 
-        for subset_name, subset in self._extractor.subsets().items():
+        subsets = self._extractor.subsets()
+        pbars = self._ctx.progress_reporter.split(len(subsets))
+        for pbar, (subset_name, subset) in zip(pbars, subsets.items()):
             task_converters = self._make_task_converters()
             for task_conv in task_converters.values():
                 task_conv.save_categories(subset)
             if CocoTask.panoptic in task_converters:
                 self._make_segmentation_dir(subset_name)
 
-            for item in subset:
-                if self._save_images:
-                    if item.has_image:
-                        self._save_image(item, subdir=osp.join(self._images_dir,
-                            '' if self._merge_images else subset_name))
-                    else:
-                        log.debug("Item '%s' has no image info", item.id)
-                for task_conv in task_converters.values():
-                    task_conv.save_image_info(item,
-                        self._make_image_filename(item))
-                    task_conv.save_annotations(item)
+            for item in pbar.iter(subset, desc=f'Exporting {subset_name}'):
+                try:
+                    if self._save_images:
+                        if item.has_image:
+                            self._save_image(item, subdir=osp.join(
+                                self._images_dir,
+                                '' if self._merge_images else subset_name))
+                        else:
+                            log.debug("Item '%s' has no image info", item.id)
+
+                    for task_conv in task_converters.values():
+                        try:
+                            task_conv.save_image_info(item,
+                                self._make_image_filename(item))
+                            task_conv.save_annotations(item)
+                        except Exception as e:
+                            self._report_annotation_error(e,
+                                item_id=(item.id, item.subset))
+                except Exception as e:
+                    self._report_item_error(e, item_id=(item.id, item.subset))
 
             for task, task_conv in task_converters.items():
                 ann_file = osp.join(self._ann_dir,