+++ datumaro/plugins/yolo_format/extractor.py
@@ -1,8 +1,9 @@
-# Copyright (C) 2019-2021 Intel Corporation
+# Copyright (C) 2019-2022 Intel Corporation
 #
 # SPDX-License-Identifier: MIT
 
 from collections import OrderedDict
+from typing import Dict
 import os.path as osp
 import re
 
@@ -39,8 +40,8 @@ class YoloExtractor(SourceExtractor):
         def categories(self):
             return self._parent.categories()
 
-    def __init__(self, config_path, image_info=None):
-        super().__init__()
+    def __init__(self, config_path, image_info=None, **kwargs):
+        super().__init__(**kwargs)
 
         if not osp.isfile(config_path):
             raise Exception("Can't read dataset descriptor file '%s'" %
@@ -96,7 +97,7 @@ class YoloExtractor(SourceExtractor):
                 )
             subsets[subset_name] = subset
 
-        self._subsets = subsets
+        self._subsets: Dict[str, YoloExtractor.Subset] = subsets
 
         self._categories = {
             AnnotationType.label:
@@ -187,8 +188,10 @@ class YoloExtractor(SourceExtractor):
         return label_categories
 
     def __iter__(self):
-        for subset in self._subsets.values():
-            for item in subset:
+        subsets = self._subsets
+        pbars = self._ctx.progress_reporter.split(len(subsets))
+        for pbar, (subset_name, subset) in zip(pbars, subsets.items()):
+            for item in pbar.iter(subset, desc=f"Parsing '{subset_name}'"):
                 yield item
 
     def __len__(self):